requirements.txt
Faker
faker-commerce
kafka-python
python-dotenv
psycopg2
